Add useColor hook
