+++ vite.config.ts
@@ -28,6 +28,7 @@ export default defineConfig({
     alias: {
       '@docs': path.resolve(__dirname, './src/docs'),
       '@components': path.resolve(__dirname, './src/components'),
+      '@hooks': path.resolve(__dirname, './src/hooks'),
     },
   }
 })
